refactor: remove search code duplication
